Parser/requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
astor==0.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
packaging==24.1
page==0.2
pandas==2.2.2
playwright==1.45.1
pluggy==1.5.0
pyee==11.1.0
pyodbc==5.1.0
pytest==8.3.2
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.5.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
requests==2.32.3
screeninfo==0.8.1
six==1.16.0
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
